app/requirements.txt
Django>=5.0,<5.2
pydantic[email]>=2.6,<3
psycopg[binary]>=3.1
PyJWT>=2.8
weasyprint>=62
python-dotenv>=1.0
requests>=2.31
python-json-logger>=2.0,<3
gunicorn>=22.0
